Add files via upload
This graph shows a bar plot that depicts the amount of countries whose happiness score had a percent decrease that was less than or equal to the percent decrease for the United States in one bar and the amount of countries whose happiness score had a percent decrease that was more than the percent decrease for the United States in the other. The colors signify the continent of each listed country.