Add files via upload
This pie chart depicts the percent decrease of happiness score for countries relative to that of the United States. The two categories are "Less" and "More" and the plot shows the percentage of countries falling into each category. "Less" shows the percent of countries whose happiness score decreased by a lesser percentage than or the same percentage as the United States. "More" shows the percent of countries whose happiness score decreased by a greater percentage than the United States.