Add files via upload
The scatter plot shows the yearly happiness score of the United States over the years 2006-2021. The x-axis shows the year and the y-axis shows the happiness score. The line on the graph is the fit from a linear regression on the United States data for the happiness score and the year.