Add files via upload
This bar plot depicts the percent decrease of happiness score for countries relative to that of the United States. One bar, labeled "Less or Equal", counts the countries whose happiness score decreased by a lesser percentage than or the same percentage as the United States. The other bar, labeled "More", counts the countries whose happiness score decreased by a greater percentage than the United States. The x-axis shows which bar represents which group and the y-axis shows the number of countries in each category. The colors on the bars relate to the continents of the counted countries.